Add annotation for build failures
