+++ scripts/utils.mjs
@@ -2549,6 +2549,48 @@ export function parseAnnotations(content, options = {}) {
 }
 
 /**
+ * @typedef {object} BuildkiteAnnotation
+ * @property {string} [context]
+ * @property {string} label
+ * @property {string} content
+ * @property {"error" | "warning" | "info"} [style]
+ * @property {number} [priority]
+ * @property {number} [attempt]
+ */
+
+/**
+ * @param {BuildkiteAnnotation} annotation
+ */
+export function reportAnnotationToBuildKite({ context, label, content, style = "error", priority = 3, attempt = 0 }) {
+  const { error, status, signal, stderr } = spawnSync(
+    "buildkite-agent",
+    ["annotate", "--append", "--style", `${style}`, "--context", `${context || label}`, "--priority", `${priority}`],
+    {
+      input: content,
+      stdio: ["pipe", "ignore", "pipe"],
+      encoding: "utf-8",
+      timeout: spawnTimeout,
+      cwd,
+    },
+  );
+  if (status === 0) {
+    return;
+  }
+  if (attempt > 0) {
+    const cause = error ?? signal ?? `code ${status}`;
+    throw new Error(`Failed to create annotation: ${label}`, { cause });
+  }
+  const buildLabel = getTestLabel();
+  const buildUrl = getBuildUrl();
+  const platform = buildUrl ? `<a href="${buildUrl}">${buildLabel}</a>` : buildLabel;
+  let errorMessage = `<details><summary><code>${label}</code> - annotation error on ${platform}</summary>`;
+  if (stderr) {
+    errorMessage += `\n\n\`\`\`terminal\n${escapeCodeBlock(stderr)}\n\`\`\`\n\n</details>\n\n`;
+  }
+  reportAnnotationToBuildKite({ label: `${label}-error`, content: errorMessage, attempt: attempt + 1 });
+}
+
+/**
  * @param {object} obj
  * @param {number} indent
  * @returns {string}
